Fix: Refactor to address Pull Request comments
