+++ tests/data/specification-latest/datatype.csv
